# Used shared backing array for Msg.Copy